Rewrite setPrevUrl function
