+++ src/components/Category/CategoryExplorer.js
@@ -18,6 +18,7 @@ import { useMediaQuery } from 'react-responsive';
 import MapPlaceholder from "../MapPlaceholder/MapPlaceholder";
 import { light } from "../../settings";
 import { lighten } from 'polished';
+import { handleSetPrevUrl } from "../../util/functions/handleSetPrevUrl";
 
 export const FILTER_MODIFIER = {
     grey: () => `
@@ -53,9 +54,6 @@ const CategoryExplorer = ({ category, onClick }) => {
   const {prevUrl, setPrevUrl} = useContext(PrevUrlContext);
   const {prevUrlParams, setPrevUrlParams} = useContext(PrevUrlParamsContext);
   const {mapToggle, setMapToggle} = useContext(MapToggleContext);
-  const paramsArray = ["category_explorer", "postcode", "service_search", "support_service", "categories", "demographic"];
-  const currentSearch = window.location.search;
-  let paramObj = {};
   const [showMap, setShowMap] = useState("false");
   const [fetchOnce, setfetchOnce] = useState(false);
 
@@ -67,17 +65,6 @@ const CategoryExplorer = ({ category, onClick }) => {
   const Mobile = ({ children }) => {
     const isMobile = useMediaQuery({ maxWidth: 767 })
     return isMobile ? children : null
-  }
-
-  function createParamObj(currentSearch, paramsArray) {
-    const queryParts = currentSearch.substring(1).split(/[&;]/g);
-    const arrayLength = queryParts.length;
-    for (let i = 0; i < arrayLength; i++) {
-      const queryKeyValue = queryParts[i].split("=");
-      if (paramsArray.includes(queryKeyValue[0])) {
-        paramObj[queryKeyValue[0]] = queryKeyValue[1];
-      }
-    }
   }
   
   useEffect(() => {
@@ -108,20 +95,12 @@ const CategoryExplorer = ({ category, onClick }) => {
       setfetchOnce(true);
     }
     
-    if (prevUrl.length == 0 && prevUrlParams.length == 0) {
-      let prevUrlArray = [""];
-      let prevUrlParamsArray = [{}];
-
-      // setPrevUrl
-      if (currentSearch) {
-        prevUrlArray.push(currentSearch);
-        setPrevUrl(prevUrlArray);
-      }
-
-      // setPrevUrlParams
-      createParamObj(currentSearch, paramsArray);
-      prevUrlParamsArray.push(paramObj);
-      setPrevUrlParams(prevUrlParamsArray);
+    const setPrevUrlVals = handleSetPrevUrl({
+      prevUrl, prevUrlParams
+    });
+    if (setPrevUrlVals) {
+      setPrevUrl(setPrevUrlVals.prevUrlArray);
+      setPrevUrlParams(setPrevUrlVals.prevUrlParamsArray);
     }
 
     if (mapToggle === "true") {
